Changed .html to .rst to fix link
Fixed the link at the beginning of the page that links to Chapter 3 and at the end of the page that links to Chapter 9. The links are actually jacobian/djangobook.com/blob/master/chapter03.rst and jacobian/djangobook.com/blob/master/chapter09.rst not ../chapter03.html and ../chapter09.html, respectively.